Function Development And Improved

+++ services/gateway_manager/src/main/java/com/futurewei/alcor/gatewaymanager/service/GatewayService.java
@@ -7,8 +7,7 @@ import com.futurewei.alcor.gatewaymanager.dao.GWAttachmentRepository;
 import com.futurewei.alcor.gatewaymanager.dao.GatewayRepository;
 import com.futurewei.alcor.gatewaymanager.entity.GWAttachment;
 import com.futurewei.alcor.gatewaymanager.entity.ResourceType;
-import com.futurewei.alcor.web.entity.gateway.VpcInfo;
-import com.futurewei.alcor.web.entity.gateway.VpcInfoSub;
+import com.futurewei.alcor.gatewaymanager.entity.StatusEnum;
 import com.futurewei.alcor.web.entity.gateway.*;
 import com.futurewei.alcor.web.restclient.GatewayManagerRestClinet;
 import lombok.extern.slf4j.Slf4j;
@@ -16,16 +15,17 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.retry.annotation.EnableRetry;
 import org.springframework.retry.annotation.Recover;
 import org.springframework.retry.annotation.Retryable;
+import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 import java.util.concurrent.CompletionException;
+import java.util.function.Supplier;
 
 @Slf4j
 @EnableRetry
 @Service
+@Component("com.futurewei.alcor.web.restclient")
 public class GatewayService {
 
     @Autowired
@@ -37,10 +37,15 @@ public class GatewayService {
     @Autowired
     private GatewayManagerRestClinet restClinet;
 
+    private static final String GATEWAY_DESCRIPTION = "internal gateway";
+    private static final String GATEWAY_SUFFIX_NAME = " name";
+    private static final String ATTACHMENT_NAME_PREFIX = "attachment - ";
+
+
     public GatewayService() {
     }
 
-    public void createGatewayInfo(String projectId, VpcInfo vpcInfo) throws CacheException {
+    public GatewayInfo createGatewayInfo(String projectId, VpcInfo vpcInfo) throws Exception {
 
         // check whether the project type is zeta
 
@@ -60,13 +65,27 @@ public class GatewayService {
             log.info("failed to create vpc in the Gateway, Exception detail: {}", e.getMessage());
             rollback(executor, gatewayInfo, gatewayEntity, projectId);
         }
-        updateDPMCacheGateway(executor, gatewayInfo, gatewayEntity, projectId);
+        executor.runAsync((Supplier<Void>) () -> {
+            try {
+                updateDPMCacheGateway(executor, gatewayInfo, gatewayEntity, projectId);
+            } catch (Exception e) {
+                log.info("update GatewayEntity status to READY failed, detail message: {}", e.getMessage());
+                //TODO how to handle this exception? rollback all?
+            }
+            return null;
+        });
+        return gatewayInfo;
     }
 
-    //TODO: need to supplemented
-    public void updateGatewayInfoForZeta(String projectId, GatewayInfo newGatewayInfo) throws CacheException {
+    public void updateGatewayInfoForZeta(String projectId, GatewayInfo newGatewayInfo) throws Exception {
         GatewayInfo oldGatewayInfo = gatewayRepository.findItem(newGatewayInfo.getResourceId());
+        if (oldGatewayInfo == null) {
+            throw new Exception(ExceptionMsgConfig.GATEWAYINFO_NOT_FOUND.getMsg());
+        }
         List<GatewayEntity> gatewayEntities = oldGatewayInfo.getGatewayEntities();
+        if (gatewayEntities == null || gatewayEntities.size() == 0) {
+            throw new Exception(ExceptionMsgConfig.GATEWAYS_IS_NULL.getMsg());
+        }
 
         Map<String, GWAttachment> attachmentsMap = gwAttachmentRepository.findAllItems();
         for (GatewayEntity newGatewayEntity : newGatewayInfo.getGatewayEntities()) {
@@ -74,7 +93,7 @@ public class GatewayService {
                 if (attachment.getResourceId().equals(newGatewayInfo.getResourceId())) {
                     for (GatewayEntity oldGatewayEntity : gatewayEntities) {
                         if (oldGatewayEntity.getId().equals(attachment.getGatewayId()) && oldGatewayEntity.getType().equals(newGatewayEntity.getType())) {
-                            oldGatewayEntity.setState(newGatewayEntity.getState());
+                            oldGatewayEntity.setStatus(newGatewayEntity.getStatus());
                         }
                     }
                 }
@@ -86,23 +105,31 @@ public class GatewayService {
 
     public void deleteGatewayInfoForZeta(String projectId, String vpcId) throws Exception {
         GatewayInfo gatewayInfo = gatewayRepository.findItem(vpcId);
+        if (gatewayInfo == null) {
+            throw new Exception(ExceptionMsgConfig.GATEWAYINFO_NOT_FOUND.getMsg());
+        }
         Map<String, GWAttachment> attachmentsMap = gwAttachmentRepository.findAllItems();
-        gatewayRepository.deleteGatewayInfoForZeta(vpcId,gatewayInfo,attachmentsMap);
+        gatewayRepository.deleteGatewayInfoForZeta(vpcId, gatewayInfo, attachmentsMap);
     }
 
 
     private void createGatewayAndAttachment(VpcInfo vpcInfo, GatewayInfo gatewayInfo, GatewayEntity gatewayEntity) throws CacheException {
+        gatewayEntity.setId(UUID.randomUUID().toString());
+        gatewayEntity.setDescription(GATEWAY_DESCRIPTION);
         gatewayEntity.setType(GatewayType.ZETA);
-        gatewayEntity.setName(gatewayEntity.getType() + " name");
-        gatewayEntity.setState("PENDING");
+        gatewayEntity.setName(gatewayEntity.getType() + GATEWAY_SUFFIX_NAME);
+        gatewayEntity.setStatus(StatusEnum.PENDING.getStatus());
         gatewayInfo.setResourceId(vpcInfo.getVpcId());
-        List<GatewayEntity> gatewayEntities = Collections.singletonList(gatewayEntity);
+        ArrayList<GatewayEntity> gatewayEntities = new ArrayList<>();
+        gatewayEntities.add(gatewayEntity);
         gatewayInfo.setGatewayEntities(gatewayEntities);
-        gatewayInfo.setStatus("available");
+        gatewayInfo.setStatus(StatusEnum.AVAILABLE.getStatus());
         // create attachment and attach it to the GatewayEntity
-        GWAttachment attachment = new GWAttachment(GatewayType.ZETA.getGatewayType() + "attachment - " + ResourceType.VPC.name(),
-                ResourceType.VPC, vpcInfo.getVpcId(), gatewayEntity.getId(), "available", vpcInfo.getVpcVni());
-        gatewayEntity.setAttachments(Collections.singletonList(attachment.getId()));
+        GWAttachment attachment = new GWAttachment(GatewayType.ZETA.getGatewayType() + ATTACHMENT_NAME_PREFIX + ResourceType.VPC.name(),
+                ResourceType.VPC, vpcInfo.getVpcId(), gatewayEntity.getId(), StatusEnum.AVAILABLE.getStatus(), vpcInfo.getVpcVni());
+        ArrayList<String> attachmentIds = new ArrayList<>();
+        attachmentIds.add(attachment.getId());
+        gatewayEntity.setAttachments(attachmentIds);
 
         gwAttachmentRepository.addItem(attachment);
     }
@@ -127,16 +154,18 @@ public class GatewayService {
         return gatewayInfo.getGatewayEntities();
     }
 
-    private void updateDPMCacheGateway(AsyncExecutor executor, GatewayInfo gatewayInfo, GatewayEntity gatewayEntity, String projectId) {
+    private void updateDPMCacheGateway(AsyncExecutor executor, GatewayInfo gatewayInfo, GatewayEntity gatewayEntity, String projectId) throws Exception {
         List<Object> result = executor.joinAll();
+        log.info("wait for all future result, result: {}", result);
         if (result.size() > 0) {
             for (Object o : result) {
                 if (o.getClass().equals(GatewayIpJson.class)) {
                     GatewayIpJson gatewayIpJson = (GatewayIpJson) o;
                     List<GatewayIp> gatewayIps = gatewayIpJson.getGatewayIps();
                     gatewayEntity.setIps(gatewayIps);
-                    gatewayEntity.setState("READY");
+                    gatewayEntity.setStatus(StatusEnum.READY.getStatus());
                     restClinet.updateDPMCacheGateway(projectId, gatewayInfo);
+                    gatewayRepository.addItem(gatewayInfo);
                 }
             }
         }
@@ -150,7 +179,7 @@ public class GatewayService {
 
 
     @Recover
-    private void workAfterRetryFailed(AsyncExecutor executor, GatewayInfo gatewayInfo, GatewayEntity gatewayEntity, String projectId) {
+    private void workAfterRetryFailed(AsyncExecutor executor, GatewayInfo gatewayInfo, GatewayEntity gatewayEntity, String projectId) throws Exception {
         rollback(executor, gatewayInfo, gatewayEntity, projectId);
     }
 
@@ -160,18 +189,20 @@ public class GatewayService {
      * @param executor
      * @param projectId
      */
-    private void rollback(AsyncExecutor executor, GatewayInfo gatewayInfo, GatewayEntity gatewayEntity, String projectId) {
+    private void rollback(AsyncExecutor executor, GatewayInfo gatewayInfo, GatewayEntity gatewayEntity, String projectId) throws Exception {
         List<Object> result = executor.joinAll();
-        log.info("start rollback and update the state for GatewayEntity, the executor's result is: {}", result);
-        for (Object ob : result) {
+        log.info("start rollback and update the status for GatewayEntity, the executor's result is: {}", result);
+        Iterator<Object> iterator = result.iterator();
+        while (iterator.hasNext()) {
+            Object ob = iterator.next();
             if (ob.getClass().equals(GatewayIpJson.class)) {
                 restClinet.deleteVPCInZetaGateway(((GatewayIpJson) ob).getVpcId());
-                result.remove(ob);
+                iterator.remove();
             }
             if (ob.getClass().equals(String.class)) {
-                gatewayEntity.setState("FAILED");
+                gatewayEntity.setStatus(StatusEnum.FAILED.getStatus());
                 restClinet.updateDPMCacheGateway(projectId, gatewayInfo);
-                result.remove(ob);
+                iterator.remove();
             }
         }
     }
